hackathon1thang6/requirements.txt

apache-airflow-providers-postgres
apache-airflow-providers-amazon


pandas>=2.0.0
numpy>=1.24.0


psycopg2-binary>=2.9.0


boto3>=1.26.0
botocore>=1.29.0


python-dotenv>=1.0.0


requests>=2.31.0


pydantic>=2.0.0

python-dateutil>=2.8.2 